feat: add highlight and cmd+click to helm value files
